Restore expected TypeError.
I must have removed this to get rid of a error with some test sample
during the testing of the branch, but it's absolutely needed so that
mime attachments get shown properly. If the TypeError raises
inapropiately due to some malformed part_map, then we will have to
catch it using a workaround somewhere else.