[bug] fix the rendering of nested multipart
This commit fix a very simplistic and until now broken handling of
nested multipart that went undetected due to the structure of the mails
used in tests until now.  Incidentally, the way that Mail.app structures
attachments made this bug noticeable.

There was also an off-by-one indexing error when retrieving the subpart
message for a given subpart.

Be aware that the current implementation will only handle correctly 2
levels of multipart nesting. Extending beyond in a more generic way will
need further work.

Closes: #7244