[bug] workaround for off-by-one error on nested multipart
For some reason that I haven't discovered yet, nested multipart is
hitting an off-by-one error (that had been wrongly ammended in a
previous commit, breaking many other cases but fixing the particular
Mail.app sample I was working with). This is just a temporary hack
to make all the current regression tests happy, but further
investigation is needed to discover the cause of the off-by-one part
retrieval solved and correctly documented.