[bug] Allow pixelated integration not to interfere with thunderbird
One of the pixelated adaptors was trying to access a non-existing
attribute in HashableMailbox, which for some reason was blocking the
operation of the imap server (uncatched exception in listeners call
maybe).

adding an attribute skips this error and therefore allows seamless use
of both pixelated and thunderbird user agents at the same time.

Resolves: #8083