[feat] Remove senders X-Leap-* headers if the email came with them
We use the X-Leap-Signature and X-Leap-Encryption to signal the
signature and encryption status of emails. An attacker could add this
headers and trick bitmask to believe that the email was signed and/or
encrypted. Now we remove this headers from the original email if they
are present before adding ours.

- Resolves: #7429